requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0

# Date and time
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: For production deployment
gunicorn==21.2.0

# Audio and speech processing
twilio
speechrecognition
pygame
pydub
elevenlabs

# AI and ML
ollama

# Data processing
pandas
openpyxl

# Utilities
pytz
apscheduler
psutil